Create CPP Basic Algo
(Issue: Logical Error not yet done)
Suppose that Billing Amount is a double variable, which denotes the amount you need to pay to department store.
If you pay the full amount, you get 1% of the billing amount, as a credit on your next bill;
if you pay at least 50% of the billing amount, the penalty is 5% of the balance;
if you pay at least 20% of the billing amount and less than 50% of the billing amount, the penalty is 10% of the balance;
otherwise, the penalty is 20% of the balance.

Design an algorithm that prompts the user to enter the billing amount and the desired payment.
Write a program that calculates and outputs the credit or the remaining balance. if the amount is not paid in full, the program should also output the penalty amount.